+++ test/proxy/ERC1967/ERC1967Proxy.test.js
@@ -3,11 +3,9 @@ const shouldBehaveLikeProxy = require('../Proxy.behaviour');
 const ERC1967Proxy = artifacts.require('ERC1967Proxy');
 
 contract('ERC1967Proxy', function (accounts) {
-  const [proxyAdminOwner] = accounts;
-
-  const createProxy = async function (implementation, _admin, initData, opts) {
+  const createProxy = async function (implementation, _owner, initData, opts) {
     return ERC1967Proxy.new(implementation, initData, opts);
   };
 
-  shouldBehaveLikeProxy(createProxy, proxyAdminOwner);
+  shouldBehaveLikeProxy(createProxy, accounts);
 });
