+++ test/utils/Create2.test.js
@@ -1,5 +1,5 @@
 const { balance, ether, expectEvent, expectRevert, send } = require('@openzeppelin/test-helpers');
-const { computeCreate2Address } = require('../helpers/create2');
+const { computeCreate2Address } = require('../helpers/create');
 const { expect } = require('chai');
 const { expectRevertCustomError } = require('../helpers/customError');
